requirements.txt
streamlit
openai
pandas
numpy
python-dotenv
PyPDF2
pdfplumber
PyMuPDF
pdf2image
pytesseract
Pillow
requests
plotly
scikit-learn
transformers
nltk
spacy
huggingface-hub
matplotlib
seaborn
beautifulsoup4
opencv-python-headless
